requirements.txt
# LangChain 核心
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LLM Providers
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# 工具生态
langchain-tools>=0.1.0
duckduckgo-search>=5.0

# TTS
edge-tts>=6.0.0

# Web 服务
flask>=3.0.0
flask-cors>=4.0.0

# 环境变量
python-dotenv>=1.0.0

# 工具辅助
requests>=2.31.0
aiohttp>=3.9.0
